+++ src/actions/index.js
@@ -567,6 +567,8 @@ export const DFR3CURVE_LOADING = "DFR3CURVE_LOADING";
 export const DFR3CURVE_LOAD_COMPLETE = "DFR3CURVE_LOAD_COMPLETE";
 export const DFR3MAPPING_LOADING = "DFR3MAPPING_LOADING";
 export const DFR3MAPPING_LOAD_COMPLETE = "DFR3MAPPING_LOAD_COMPLETE";
+export const SEMANTIC_LOADING = "SEMANTIC_LOADING";
+export const SEMANTIC_LOAD_COMPLETE = "SEMANTIC_LOAD_COMPLETE";
 
 export function loading(component) {
 	return (dispatch: Dispatch) => {
@@ -639,6 +641,65 @@ export function getOutputDataset(executionId: String) {
 	};
 }
 
+// Semantic Functions
+export const RECEIVE_SEMANTICS = "RECEIVE_SEMANTICS";
+
+export function receiveSemantics(type: string, json) {
+	return (dispatch: Dispatch) => {
+		dispatch({
+			type: type,
+			semantics: json,
+			receivedAt: Date.now()
+		});
+	};
+}
+
+// TODO - WIP
+export function fetchSemantics(space, limit, offset) {
+	let endpoint = `${config.semanticServiceType}?limit=${limit}&skip=${offset}&detail=true`;
+
+	if (space !== null && space !== "All") {
+		endpoint = `${endpoint}&space=${space}`;
+	}
+
+	return (dispatch: Dispatch) => {
+		dispatch(loading(SEMANTIC_LOADING));
+		return fetch(endpoint, { mode: "cors", headers: getHeader() }).then((response) => {
+			dispatch(loadComplete(SEMANTIC_LOAD_COMPLETE));
+			if (response.status === 200) {
+				response.json().then((json) => {
+					dispatch(receiveSemantics(RECEIVE_SEMANTICS, json));
+				});
+			} else if (response.status === 401) {
+				cookies.remove("Authorization");
+				dispatch(receiveSemantics(LOGIN_ERROR, []));
+			} else {
+				dispatch(receiveSemantics(RECEIVE_SEMANTICS, []));
+			}
+		});
+	};
+}
+
+export function searchSemantics(keyword, limit, offset) {
+	let endpoint = `${config.semanticServiceType}/search?text=${keyword}&limit=${limit}&skip=${offset}`;
+	return (dispatch: Dispatch) =>{
+		dispatch(loading(SEMANTIC_LOADING));
+		return fetch(endpoint, { mode: "cors", headers: getHeader() }).then((response) => {
+			dispatch(loadComplete(SEMANTIC_LOAD_COMPLETE));
+			if (response.status === 200) {
+				response.json().then((json) => {
+					dispatch(receiveSemantics(RECEIVE_SEMANTICS, json));
+				});
+			} else if (response.status === 401) {
+				cookies.remove("Authorization");
+				dispatch(receiveSemantics(LOGIN_ERROR, []));
+			} else {
+				dispatch(receiveSemantics(RECEIVE_SEMANTICS, []));
+			}
+		});
+	};
+}
+
 export async function executeDatawolfWorkflowHelper(workflowid, creatorid, title, description, parameters, datasets) {
 	const datawolfUrl = `${config.dataWolf}executions`;
 	const dataToSubmit = {
@@ -706,7 +767,24 @@ function getDatawolfHeader() {
 }
 
 export async function getRepoVersion() {
-	const versionRequest = await fetch("tags/github.json");
+	const versionRequest = await fetch("https://raw.githubusercontent.com/IN-CORE/IN-CORE/main/tags.json", {
+		method: "GET",
+		redirect: "follow"
+	});
+
+	const githubVersionResponseFallback = {
+		"in-core": "NA",
+		"incore-auth": "NA",
+		"incore-docs": "NA",
+		"incore-helm": "NA",
+		"incore-lab": "NA",
+		"incore-services": "NA",
+		"incore-ui": "NA",
+		"plotting-service": "NA",
+		"pyincore": "NA",
+		"pyincore-viz": "NA",
+		"pyincore-data": "NA"
+	};
 
 	try {
 		const githubVersions = await versionRequest.json();
@@ -714,6 +792,7 @@ export async function getRepoVersion() {
 	} catch (error) {
 		// if fail just log the erorr and return null
 		console.log(error);
-		return null;
+		// return default valued expected object.
+		return githubVersionResponseFallback;
 	}
 }