+++ src/components/App.jsx
@@ -287,13 +287,13 @@ class App extends Component {
 							variant="determinate"
 							className={classes.customProgressBar}
 							value={
-								(this.props.usage["total_file_size_of_datasets_byte"] /
+								((this.props.usage["total_file_size_of_datasets_byte"] ?? 0) /
 									this.props.allocations["total_file_size_of_datasets_byte"]) *
 								100
 							}
 						/>
 						<Typography className={classes.fontLight} style={{ fontSize: "10px" }}>
-							Data {this.props.usage["total_file_size_of_datasets"]} of{" "}
+							Data {this.props.usage["total_file_size_of_datasets"] ?? 0} of{" "}
 							{this.props.allocations["total_file_size_of_datasets"]} used
 						</Typography>
 					</Box>
@@ -302,14 +302,29 @@ class App extends Component {
 							variant="determinate"
 							className={classes.customProgressBar}
 							value={
-								(this.props.usage["total_file_size_of_hazard_datasets_byte"] /
+								((this.props.usage["total_file_size_of_hazard_datasets_byte"] ?? 0) /
 									this.props.allocations["total_file_size_of_hazard_datasets_byte"]) *
 								100
 							}
 						/>
 						<Typography className={classes.fontLight} style={{ fontSize: "10px" }}>
-							Hazard {this.props.usage["total_file_size_of_hazard_datasets"]} of{" "}
+							Hazard {this.props.usage["total_file_size_of_hazard_datasets"] ?? 0} of{" "}
 							{this.props.allocations["total_file_size_of_hazard_datasets"]} used
+						</Typography>
+					</Box>
+					<Box className={classes.status}>
+						<LinearProgress
+							variant="determinate"
+							className={classes.customProgressBar}
+							value={
+								((this.props.usage["total_number_of_dfr3"] ?? 0) /
+									this.props.allocations["total_number_of_dfr3"]) *
+								100
+							}
+						/>
+						<Typography className={classes.fontLight} style={{ fontSize: "10px" }}>
+							DFR3 {this.props.usage["total_number_of_dfr3"] ?? 0} of{" "}
+							{this.props.allocations["total_number_of_dfr3"]} used
 						</Typography>
 					</Box>
 					<Divider orientation="horizontal" />
@@ -330,6 +345,16 @@ class App extends Component {
 					>
 						<Link href={`mailto:${config.mailingList}`} target="_blank" style={{ textDecoration: "none" }}>
 							Contact Us
+						</Link>
+					</MenuItem>
+					<MenuItem
+						className={classes.denseStyle}
+						onClick={() => {
+							this.handleProfileMenuClose();
+						}}
+					>
+						<Link href={config.slackInvitationLink} target="_blank" style={{ textDecoration: "none" }}>
+							Join Slack
 						</Link>
 					</MenuItem>
 					<MenuItem
@@ -405,6 +430,16 @@ class App extends Component {
 					}}
 				>
 					Hazard Viewer
+				</MenuItem>
+				<MenuItem
+					className={classes.denseStyle}
+					onClick={() => {
+						this.handleViewerMenuClose();
+						browserHistory.push("/SemanticViewer");
+						fetch("/SemanticViewer");
+					}}
+				>
+					Dataset Type Viewer
 				</MenuItem>
 			</Menu>
 		);