+++ src/components/HazardViewer.jsx
@@ -80,9 +80,6 @@ const styles = {
 		display: "inline-block",
 		margin: "auto 5px"
 	},
-	hide: {
-		display: "none"
-	},
 	paperFooter: {
 		padding: theme.spacing(2),
 		borderTop: "1px solid #eeeeee",
@@ -645,70 +642,74 @@ class HazardViewer extends Component {
 							</Grid>
 
 							{/* Metadata */}
-							<Grid
-								item
-								lg={8}
-								md={8}
-								xl={8}
-								xs={12}
-								className={this.state.selectedHazard ? null : classes.hide}
-							>
-								<Paper variant="outlined" className={classes.main}>
-									<IconButton
-										aria-label="Close"
-										onClick={this.closeMetadata}
-										className={classes.metadataCloseButton}
+							{
+								this.state.metadataClosed ?
+									<></>
+									:
+									<Grid
+										item
+										lg={8}
+										md={8}
+										xl={8}
+										xs={12}
 									>
-										<CloseIcon fontSize="small" />
-									</IconButton>
-									{Object.keys(selected_hazard_detail).length > 0 ? (
-										<div>
-											<div className={classes.paperHeader}>
-												<Typography variant="subtitle1">Metadata</Typography>
-											</div>
-											<div className={classes.metadata}>
-												<Button
-													color="primary"
-													variant="contained"
-													className={classes.inlineButtons}
-													size="small"
-													onClick={this.exportJson}
-												>
-													Download Metadata
-												</Button>
-												<CopyToClipboard text={this.state.selectedHazard.id}>
-													<Button
-														color="secondary"
-														variant="contained"
-														className={classes.inlineButtons}
-														size="small"
-													>
-														Copy ID
-													</Button>
-												</CopyToClipboard>
-												<Button
-													color="secondary"
-													variant="contained"
-													className={classes.inlineButtons}
-													size="small"
-													onClick={this.onClickDelete}
-												>
-													DELETE
-												</Button>
-											</div>
-											<div className={classes.metadata}>
-												<NestedInfoTable
-													data={selected_hazard_detail}
-													selectedHazardDataset={this.state.selectedHazardDatasetId}
-													onClick={this.preview}
-												/>
-											</div>
-										</div>
-									) : (
-										<div />
-									)}
-								</Paper>
-							</Grid>
+										<Paper variant="outlined" className={classes.main}>
+											<IconButton
+												aria-label="Close"
+												onClick={this.closeMetadata}
+												className={classes.metadataCloseButton}
+											>
+												<CloseIcon fontSize="small" />
+											</IconButton>
+											{Object.keys(selected_hazard_detail).length > 0 ? (
+												<div>
+													<div className={classes.paperHeader}>
+														<Typography variant="subtitle1">Metadata</Typography>
+													</div>
+													<div className={classes.metadata}>
+														<Button
+															color="primary"
+															variant="contained"
+															className={classes.inlineButtons}
+															size="small"
+															onClick={this.exportJson}
+														>
+															Download Metadata
+														</Button>
+														<CopyToClipboard text={this.state.selectedHazard.id}>
+															<Button
+																color="secondary"
+																variant="contained"
+																className={classes.inlineButtons}
+																size="small"
+															>
+																Copy ID
+															</Button>
+														</CopyToClipboard>
+														<Button
+															color="secondary"
+															variant="contained"
+															className={classes.inlineButtons}
+															size="small"
+															onClick={this.onClickDelete}
+														>
+															DELETE
+														</Button>
+													</div>
+													<div className={classes.metadata}>
+														<NestedInfoTable
+															data={selected_hazard_detail}
+															selectedHazardDataset={this.state.selectedHazardDatasetId}
+															onClick={this.preview}
+														/>
+													</div>
+												</div>
+											) : (
+												<div />
+											)}
+										</Paper>
+									</Grid>
+							}
 						</Grid>
 						<Version />
 					</div>