+++ CHANGELOG.md
@@ -5,8 +5,47 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
 
+## [1.9.0] - 2023-11-09
+
+### Added
+- Links to Slack channel and mailing list to the landing page [#160](https://github.com/IN-CORE/incore-ui/issues/160)
+
+### Changed
+- Missing Filters title for both Data viewer and Semantics Viewer [#156](https://github.com/IN-CORE/incore-ui/issues/156)
+- Semantic viewer needs to show description [#157](https://github.com/IN-CORE/incore-ui/issues/157)
+
+
+## [1.8.0] - 2023-10-11
+
+### Changed
+- Read IN-CORE lab quota from service endpoint [#141](https://github.com/IN-CORE/incore-ui/issues/141)
+- Change Semantic type api call logic, get complete details at first load instead of fetching on the fly. [#144](https://github.com/IN-CORE/incore-ui/issues/144)
+- Add a multi-hazard type Earthquake+Tsunami in DFR3 Viewer [#147](https://github.com/IN-CORE/incore-ui/issues/147)
+
+### Added
+- Display DFR3 quota [#137](https://github.com/IN-CORE/incore-ui/issues/137)
+- Add Semantic template link in Semantics viewer [#143](https://github.com/IN-CORE/incore-ui/issues/143)
+
+
+## [1.7.0] - 2023-08-16
+
+### Added
+- Added a viewer for Semantics [#122](https://github.com/IN-CORE/incore-ui/issues/122)
+- Added clickable link in DataViewer to show semantic definition for datatype [#123](https://github.com/IN-CORE/incore-ui/pull/123)
+
+### Fixed
+- When-user-has-no-usage-display-zero-instead-of-undefined [#110](https://github.com/IN-CORE/incore-ui/issues/110)
+- Github action to grab correct version when merged to main. [#115](https://github.com/IN-CORE/incore-ui/issues/115)
+- Improper functionality when close icon is clicked [#124](https://github.com/IN-CORE/incore-ui/issues/124)
+
+
+## [1.6.0] - 2023-04-25
+### Changed
+- Change the getVersionTags scripts to get tags from the central IN-CORE repository.[#104](https://github.com/IN-CORE/incore-ui/issues/104)
+
+
+## [1.5.0] - 2023-03-15
 ### Changed
 - builds are done on amd64 and then artifact copied to different platforms.